requirements/base.txt
anthropic==0.43.0
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
django-environ==0.12.1
django-filter==25.2
django-vite==3.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
pillow-heif==0.18.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
pypdf==5.1.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.11.0
